undo some random gtk.c changes

+++ src/common/darktable.h
@@ -402,7 +402,7 @@ void dt_show_times_f(const dt_times_t *start, const char *prefix, const char *su
 /** \brief check if file is a supported image */
 gboolean dt_supported_image(const gchar *filename);
 
-static inline size_t dt_get_num_threads()
+static inline int dt_get_num_threads()
 {
 #ifdef _OPENMP
   return omp_get_num_procs();
